Reduce Args, no numbering, write Prefix to status.yml
Only programmeName and programmePrefix are required now - programmeTitle is derived from programmeName by converting all - & _ to spaces by default.

Removed support for numbering the PROGRAMMES by default, and left this for the User to define the name with any numbering as they wish.

The Prefix (which is used to identify any projects that are primarily under this programme) is stored in status.yml, and is retrieved every time a new project is created in this programme, via `createProjectDoc.R`