Write Project Data to status.yml
When a Project is created, its FULL Name is written to status.yml (under "PROJECTS"), and includes data on the Primary programmeName, and the programmePrefix & projectIndex used in the Full project name.

Projects are added to the PROJECTS Section in Order of Creation, which can be used to sort them for Priority Lists.